added login/logout/registratiion functionality

+++ gamesquad/base/views.py
@@ -1,7 +1,11 @@
 from django.shortcuts import render, redirect
 from django.db.models import Q
-from django.contrib.auth.models import User
+from django.http import HttpResponse
 from django.contrib import messages
+from django.contrib.auth import authenticate, login, logout
+from django.contrib.auth.forms import UserCreationForm
+from django.contrib.auth.models import User
+from django.contrib.auth.decorators import login_required
 from .models import Room, RoomGame
 from .forms import RoomForm
 # Create your views here.
@@ -13,15 +17,51 @@ from .forms import RoomForm
 # ]
 
 def loginView(request):
+    page = 'login'
+    if request.user.is_authenticated:
+        return redirect('home')
+
     if request.method == 'POST':
-        username = request.POST.get('username')
+        username = request.POST.get('username').lower()
         password = request.POST.get('password')
 
         try:
             user = User.objects.get(username = username)
         except:
             messages.error(request, 'User does not exist')
-    context = {}
+
+        user = authenticate(request, username=username, password=password)
+        if user is not None:
+            login(request, user)
+            return redirect('home')
+        else:
+            messages.error(request, 'Username or password incorrect')
+
+
+    context = {'page': page}
+    return render(request, 'base/login_register.html', context)
+
+def logoutView(request):
+    logout(request)
+    return redirect('home')
+
+def registerView(request):
+    page = 'register'
+    form = UserCreationForm()
+
+    if request.method == 'POST':
+        form = UserCreationForm(request.POST)
+        if form.is_valid():
+            user = form.save(commit=False)
+            user.username = user.username.lower()
+            user.save()
+            login(request, user)
+            return redirect('home')
+        else:
+            messages.error(request, 'Invalid registration')
+
+
+    context = {'page': page, 'form': form}
     return render(request, 'base/login_register.html', context)
 
 def home(request):
@@ -42,6 +82,7 @@ def room(request, pk):
     context = {'room' : room}
     return render(request, 'base/room.html', context)
 
+@login_required(login_url='login')
 def createRoom(request):
     form = RoomForm()
     if request.method == 'POST':
@@ -52,9 +93,13 @@ def createRoom(request):
     context = {'form': form}
     return render(request, 'base/room_form.html', context)
 
+@login_required(login_url='login')
 def updateRoom(request, pk):
     room = Room.objects.get(id=pk)
     form = RoomForm(instance=room)
+
+    if request.user != room.host:
+        return HttpResponse('Only the room host can edit the room')
 
     if request.method == 'POST':
         form = RoomForm(request.POST, instance = room)
@@ -67,6 +112,10 @@ def updateRoom(request, pk):
 
 def deleteRoom(request, pk):
     room = Room.objects.get(id=pk)
+
+    if request.user != room.host:
+        return HttpResponse('Only the room host can edit the room')
+
     if request.method == 'POST':
         room.delete()
         return redirect('home')
